Fix bug in the original code on receiving resize event from gnome-shell
This is a backport of the fix initially appearing in the python code:
https://github.com/HinTak/skia-python-examples/issues/3

The mouse positioning is wrong because they are recorded related to the current
SDL window size, but is being drawn to the skia canvas which was initialized
in the original / full-screen size. There is a resizing event generated
by gnome-shell as the SDL window tries to expand to fill.

Due to this resizing event, the instructional text was also drawn
off-screen too, as the offset is about 105 pixels.

Fixes #2