gtk4 example from https://groups.google.com/g/skia-discuss/c/aw0bjhri5hU
The original posting includes this comment:

"I'm running it on Linux with Wayland.
If I create an X11 window and initialize EGL manually, it works fine, but doesn't work with GTK (draws empty white area).
I think because it initializes differently, but not sure what exactly I'm doing wrong, as I'm very new to graphics programming."